+++ chapter-1/minutes-within-decade.rb
@@ -1,0 +1,19 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 1"
+puts "# File: minutes-within-decade.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts "Constants:"
+puts "Decade: 1"
+puts "Years: 10"
+puts "Days: 365"
+puts "Hours: 24"
+puts "Months: 12"
+puts "Minutes: 60"
+puts
+
+puts "Minutes in a decade:"
+puts 10*365*24*60
+puts